Initialisation
Initialisation seems to be working. Cant confirm due to odd topology but the numbers of simplified interactions make sense. Some oddity in the number of Total interaction volume numbers but I can spot a pattern if I squint so im going to ignore it for now

+++ GlobalVariables.cs
@@ -7,7 +7,7 @@ namespace Globals
         public static int Units_in_M = 1000;
         public static int seconds_per_timestep = 60;
 
-        public static decimal max_angle_initialisation = 1;
+        public static double max_angle_initialisation = 1;
         public const long M_in_Lsecond = 299792458;
         public static  Microsoft.Extensions.Logging.LogLevel Level = Microsoft.Extensions.Logging.LogLevel.Trace;
         public static ILogger log = new Logger().log;
